requirements.txt
blinker==1.8.2
bson==0.5.10
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyJWT==2.8.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==3.0.3
